More deblob functions

+++ libdeblob.sh
@@ -11,13 +11,6 @@
 #       project.
 #########################
 
-# TODO:
-# * filedel
-# * linedel
-# * strdel
-# * add case in lineadd for instances where another line was added
-# * space (x)
-
 # Turns a file and it's path into a friendly filename
 # Usage: filetize $filepath
 filetize() {
@@ -33,22 +26,86 @@ unfiletize() {
 # Replace a string in a file
 # Usage: rep $replacee $replacer $file
 rep() {
-	sed --posix 's^'$1'^'$2'^g' $3 > /tmp/$(filetize "$3")
-	diff $3 /tmp/$(filetize "$3") >> $PATCH_DIR/$(filetize "$3").patch
+	if [ -e "/tmp/$(filetize "$3")" ]
+	then
+		sed --posix 's^'$1'^'$2'^g' /tmp/$(filetize "$3") > /tmp/$(filetize "$3").tmp
+		mv /tmp/$(filetize "$3").tmp /tmp/$(filetize "$3")
+		diff $3 /tmp/$(filetize "$3") > $PATCH_DIR/$(filetize "$3").patch
+	else
+		sed --posix 's^'$1'^'$2'^g' $3 > /tmp/$(filetize "$3")
+		diff $3 /tmp/$(filetize "$3") > $PATCH_DIR/$(filetize "$3").patch
+	fi
+
+}
+
+# Delete a string in a file
+# Usage: strdel $string $file
+strdel() {
+	if [ -e "/tmp/$(filetize "$2")" ]
+	then
+		sed 's^'$1'^^' /tmp/$(filetize "$2") > /tmp/$(filetize "$2").tmp
+		mv /tmp/$(filetize "$2").tmp /tmp/$(filetize "$2")
+		diff $2 /tmp/$(filetize "$2") > $PATCH_DIR/$(filetize "$2").patch
+	else
+		sed 's^'$1'^^' $2 > /tmp/$(filetize "$2")
+		diff $2 /tmp/$(filetize "$2") > $PATCH_DIR/$(filetize "$2").patch
+	fi
 }
 
 # Inserts a new line after another
-# Usage: addline $string $newline $file
+# Usage: lineadd $string $newline $file
 lineadd() {
-	sed 's^'$1'^'$1'\n'$2'^' $3 > /tmp/$(filetize "$3")
-	diff $3 /tmp/$(filetize "$3") >> $PATCH_DIR/$(filetize "$3").patch
+	if [ -e "/tmp/$(filetize "$3")" ]
+	then
+		sed 's^'$1'^'$1'\n'$2'^' /tmp/$(filetize "$3") > /tmp/$(filetize "$3").tmp
+		mv /tmp/$(filetize "$3").tmp /tmp/$(filetize "$3")
+		diff $3 /tmp/$(filetize "$3") > $PATCH_DIR/$(filetize "$3").patch
+	else
+		sed 's^'$1'^'$1'\n'$2'^' $3 > /tmp/$(filetize "$3")
+		diff $3 /tmp/$(filetize "$3") > $PATCH_DIR/$(filetize "$3").patch
+	fi
 }
 
+# Removes a line.
+# Usage linedel $string $file
+linedel() {
+	if [ -e "/tmp/$(filetize "$3")" ]
+	then
+		grep -v "$1" /tmp/$(filetize "$2") > /tmp/$(filetize "$2").tmp
+		mv /tmp/$(filetize "$2").tmp /tmp/$(filetize "$2")
+		diff $2 /tmp/$(filetize "$2") > $PATCH_DIR/$(filetize "$2").patch
+	else
+		grep -v "$1" $2 > /tmp/$(filetize "$2").tmp
+		diff $2 /tmp/$(filetize "$2") > $PATCH_DIR/$(filetize "$2").patch
+	fi
+}
+
+# "Copies" a file
+# Usage: filedel $file
+filecp() {
+	cp $1 /tmp/ADD_$(filetize "$2")
+}
+
+# "Deletes" a file
+# Usage: filedel $file
+filedel() {
+	touch /tmp/RM_$(filetize "$1")
+}
+
+# Applies patches.
 apply() {
 	for file in $PATCH_DIR/*
 	do
-		realname=$(echo $file | sed 's^.*/^^')
-		patch $realname < $file
+		realname=$(echo $file | sed 's^.*/^^')	# Still need unfiletizing?
+
+		if echo "$file" | grep "^RM_" > /dev/null
+		then
+			rm $realname
+		elif echo "$file" | grep "^ADD_" > /dev/null
+		then
+			cp $file $realname
+		else
+			patch $realname < $file
+		fi
 	done
 }
-
